# public/openapi.json
{
  "info": {
    "description": "Internal and public HTTP API for the Changemappers platform. This spec is generated from the route handlers under src/app/api — run `npm run openapi:generate` after adding or changing a route.",
    "title": "Changemappers API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/gdpr/export": {
      "get": {
        "description": "Returns the full GDPR data export for the authenticated user as a JSON download. Rate-limited to at most twice per day.",
        "responses": {
          "200": {
            "description": "GDPR data export (JSON attachment)"
          },
          "401": {
            "description": "Unauthorized — no valid session"
          },
          "404": {
            "description": "Authenticated user not found"
          },
          "429": {
            "description": "Rate limit exceeded (max twice per day)"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Export user data (GDPR)"
      }
    },
...
  },
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:3000"
    }
  ]
}